+++ src/GRA.Controllers/ViewModel/MissionControl/Events/EventsListViewModel.cs
@@ -17,6 +17,7 @@ namespace GRA.Controllers.ViewModel.MissionControl.Events
         public string BranchName { get; set; }
         public string ProgramName { get; set; }
         public bool CanManageLocations { get; set; }
+        public bool CommunityExperience { get; set; }
 
         public IEnumerable<GRA.Domain.Model.Branch> BranchList { get; set; }
         public IEnumerable<GRA.Domain.Model.System> SystemList { get; set; }